+++ packages/core/CHANGELOG.md
@@ -13,6 +13,13 @@ No significant changes.
 ## [4.7.11] - 2024-05-13
 
 No significant changes.
+## [unreleased]
+
+### Removed
+
+- `core/platform` member `windowsPhone`, `edge`, `androidChrome`, `android`, `ie`, `ios`, `webos`, `androidFirefox`, `firefoxOS`, `blackberry`, and `tizen`
+- `core/platform` member `gesture`, `node`, `platformName`, and `unknown`
+- `core/platform` member `touch` and `touchscreen`
 
 ## [4.9.0-alpha.1] - 2024-04-09